+++ hc/front/urls.py
@@ -39,6 +39,7 @@ channel_urls = [
     path("add_trello/settings/", views.trello_settings, name="hc-trello-settings"),
     path("<uuid:code>/checks/", views.channel_checks, name="hc-channel-checks"),
     path("<uuid:code>/name/", views.update_channel_name, name="hc-channel-name"),
+    path("<uuid:code>/edit_webhook/", views.edit_webhook, name="hc-edit-webhook"),
     path("<uuid:code>/test/", views.send_test_notification, name="hc-channel-test"),
     path("<uuid:code>/remove/", views.remove_channel, name="hc-remove-channel"),
     path(